Module c_ares::flags

Flags that may be passed when initializing a Channel.

Structs

Flags
Flags that may be passed when initializing a Channel.

Constants

EDNS
Use Extension Mechanisms for DNS.

IGNTC
If a truncated response to a UDP query is received, do not fall back to TCP; simply continue on with the truncated response.

NOALIASES
Do not honor the HOSTALIASES environment variable, which normally specifies a file of hostname translations.

NOCHECKRESP
Do not discard responses with the SERVFAIL, NOTIMP, or REFUSED response code or responses whose questions don't match the questions in the request. Primarily useful for writing clients which might be used to test or debug name servers.

NORECURSE
Do not set the "recursion desired" bit on outgoing queries, so that the name server being contacted will not try to fetch the answer from other servers if it doesn't know the answer locally.

NOSEARCH
Do not use the default search domains; only query hostnames as-is or as aliases.

PRIMARY
Only query the first server in the list of servers to query.

STAYOPEN
Do not close communications sockets when the number of active queries drops to zero.

USEVC
Always use TCP queries (the "virtual circuit") instead of UDP queries. Normally, TCP is only used if a UDP query yields a truncated result.