Module c_ares::ni_flags

Flags that may be provided on a call to get_name_info()

Structs

NIFlags
Flags that may be provided on a call to get_name_info().

Constants

DCCP
The service name is to be looked up for the DCCP protocol.

LOOKUPHOST
A hostname lookup is being requested.

LOOKUPSERVICE
A service name lookup is being requested.

NAMEREQD
An error is returned if the hostname cannot be found in the DNS.

NOFQDN
Only the nodename portion of the FQDN is returned for local hosts.

NUMERICHOST
The numeric form of the hostname is returned rather than the name.

NUMERICSCOPE
The numeric form of the scope ID is returned rather than the name.

NUMERICSERV
The numeric form of the service is returned rather than the name.

SCTP
The service name is to be looked up for the SCTP protocol.

TCP
The service name is to be looked up for the TCP protocol.

UDP
The service name is to be looked up for the UDP protocol.